prehab-enterprise/requirements.txt
# --- Core Framework ---
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib
bcrypt
alembic
python-dotenv
httpx
httpx
requests
psycopg2-binary


# --- AI & Vision (Order Matters) ---
# Install headless OpenCV first to prevent the server from trying to grab the GUI version
opencv-python-headless
ultralytics

# --- Dashboard ---
streamlit
plotly
pandas
supabase